🐛 🔖 Ensure that the normal click works on android even if actions are present
In https://github.com/bhandaribhumin/cordova-plugin-local-notification-12/commit/0b6aefe0aac71cf44feb62a975516e13adfd010e,
the plugin started ignoring regular clicks when actions are present.

```
        Action[] actions = options.getActions();
        if (actions != null && actions.length > 0 ) {
          // if actions are defined, the user must click on button actions to launch the app.
          // Don't make the notification clickable in this case
          return;
        }
```

It is not clear why this was done. I have filed
https://github.com/bhandaribhumin/cordova-plugin-local-notification-12/issues/9
but not received any response

While waiting for it to be fixed in the main repo, commenting out that code in
a fork so that we can unblock out project.