⏰ Remove exact notification calls
This fixes https://github.com/e-mission/e-mission-docs/issues/1079#issuecomment-2308917271
In the upgrade to API 14, the schedule exact alarm permission is not granted by default.
Fortunately, we don't need exact alarms for our notifications. Since we are
using a fork, we just change the calls to schedule the notification using inExactAlarms.

Add a log indicating that we are using inexact alarms, for testing.

+ remove the permission
+ bump up the version number

Testing done:

- Notification is displayed without error
- Logs show

```
08-27 17:00:57.059 11959 12185 D local-notification: Next trigger at: Tue Aug 27 17:01:02 PDT 2024
08-27 17:00:57.067 11959 12185 D local-notification: Scheduling inexact alarms with intent: PendingIntent{b58ad0c: android.os.BinderProxy@cc38655}
```